refactor: changes views with Custom Hooks
